requirements.txt
# Core dependencies
langchain==0.1.5
langchain-community==0.0.13
langchain-core==0.1.12
openai==1.12.0
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.23.8  # Also known as fitz
pdf2image==1.16.3

# Vector Database
opensearch-py==2.4.2

# Data Processing & Analysis
numpy==1.26.0
pandas==2.1.0

# LangChain specific dependencies
chromadb==0.4.22
tiktoken==0.5.2
faiss-cpu==1.7.4

# Utilities
tqdm==4.66.1
colorama==0.4.6
pydantic==2.5.0

# Optional but recommended for better performance
sentence-transformers==2.2.2 